implement fully functional canvas with token drop capabilities
